web과 stress test용 Stack 분리
